:truck: added matrix transform include to the renderer + changed window size data from float to integers.

+++ CG1/src/Core/Renderer.cpp
@@ -47,7 +47,7 @@ static void glfwErrorCallback(int error, const char* description)
 }
 
 /* initializing the Renderer object. glfw / glew */
-CG::Renderer::Renderer(const std::string& windowName, float width, float height)
+CG::Renderer::Renderer(const std::string& windowName, int width, int height)
 {
     CG_LOG_INFO("Initializing OpenGL Renderer.");
